+++ src/main/java/com/dyinfotech/annualleavebackend/repository/LeaveRequestRepository.java
@@ -1,21 +1,21 @@
 package com.dyinfotech.annualleavebackend.repository;
 
-import com.dyinfotech.annualleavebackend.domain.LeaveRequest;
-import com.dyinfotech.annualleavebackend.domain.LeaveRequestStatus;
+import java.time.LocalDate;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.util.List;
+import com.dyinfotech.annualleavebackend.domain.LeaveRequest;
+import com.dyinfotech.annualleavebackend.domain.LeaveRequestStatus;
 
 public interface LeaveRequestRepository extends JpaRepository<LeaveRequest, Long> {
 
     // 승인된 요청의 사용일수 합계 (잔여 연차 계산용)
     @Query("SELECT COALESCE(SUM(lr.useDays), 0) FROM LeaveRequest lr " +
             "WHERE lr.employee.employeeId = :employeeId AND lr.status = 'APPROVED'")
-    BigDecimal sumApprovedUseDays(@Param("employeeId") Long employeeId);
+    Float sumApprovedUseDays(@Param("employeeId") Long employeeId);
 
     // 특정 상태의 내 요청 개수
     long countByEmployee_EmployeeIdAndStatus(Long employeeId, LeaveRequestStatus status);
